backend/requirements.txt
Django==4.0.5
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
django-environ==0.9.0
Pillow==9.1.1
django-cors-headers==3.13.0
